Carry memory addresses as u32 and select the extended service by range
Widen read_memory/write_memory/write_memory_verified/compare_rel_mem and the
flash-engine write path to 24-bit u32 addresses. select_extended_memory chooses
the plain A_Memory_Write (byte-identical, <=0xFFFF) or A_MemoryExtended_Write
(>0xFFFF) from the resolved absolute address, matching ETS's per-range choice.
Extended writes scale chunks to the negotiated PID_MAX_APDU (up to 228) and are
confirmed inline by 0x1FC; extended read-back verifies via 0x1FD. Plan-time
guards lifted from the 16-bit to the 24-bit ceiling. System 7 paths keep 16-bit
behaviour unchanged.

+++ crates/bussard-mgmt/src/sys7.rs
@@ -215,7 +215,9 @@ impl LsmAccess {
         match self {
             LsmAccess::MemoryMapped { control_addr, .. } => {
                 let record = wrap_memory_lsm_record(lsm, event);
-                write_memory(l4, *control_addr, &record).await
+                // System 7 addresses are always ≤16-bit, so this stays on the
+                // plain A_Memory_Write path (see `select_extended_memory`).
+                write_memory(l4, u32::from(*control_addr), &record).await
             }
             LsmAccess::Property => {
                 property_write_request(l4, lsm, crate::load::PID_LOAD_STATE_CONTROL, 1, 1, event)
@@ -237,7 +239,7 @@ impl LsmAccess {
         match self {
             LsmAccess::MemoryMapped { status_addr, .. } => {
                 let addr = status_addr.saturating_add(u16::from(lsm.saturating_sub(1)));
-                let data = read_memory(l4, addr, 1).await?;
+                let data = read_memory(l4, u32::from(addr), 1).await?;
                 let octet = data.first().copied().ok_or_else(|| {
                     WriteError::Mgmt(crate::MgmtError::MalformedResponse {
                         address: l4.target(),
